Changed alert when on-click "DON'T BE LAME!!"
Changed it so when you click "Don't be lame!!" in the upper right hand corner a pop-up is going to appear and prompt you.
It's going to ask you "What is the one and only rule?!
if you answer correct there is going to be an alert that says "You are, correct!"
Changed by: Wille2006, AKA Knuckles.